elf: Move Elf64LoadSegments to separate module
In the effort of splitting up the ELF parser code, move the
Elf64LoadSegments and other related structs to its own module.
