+++ elf/src/lib.rs
@@ -8,12 +8,11 @@
 
 #![no_std]
 
-extern crate alloc;
-
 mod addr_range;
 mod error;
 mod file_range;
 mod header;
+mod load_segments;
 mod program_header;
 mod section_header;
 mod types;
@@ -22,14 +21,15 @@ pub use addr_range::Elf64AddrRange;
 pub use error::ElfError;
 pub use file_range::Elf64FileRange;
 use header::Elf64Hdr;
+pub use load_segments::{
+    Elf64ImageLoadSegment, Elf64ImageLoadSegmentIterator, Elf64ImageLoadVaddrAllocInfo,
+    Elf64LoadSegments,
+};
 pub use program_header::{Elf64Phdr, Elf64PhdrFlags};
 pub use section_header::{Elf64Shdr, Elf64ShdrFlags};
 pub use types::*;
 
-use alloc::vec::Vec;
-use core::cmp;
 use core::ffi;
-use core::matches;
 
 /// This struct represents a parsed 64-bit ELF file. It contains information
 /// about the ELF file's header, load segments, dynamic section, and more.
@@ -566,7 +566,7 @@ impl<'a> Elf64File<'a> {
     /// This function returns an iterator that allows iterating over the ELF segments
     /// belonging to the loaded image. It takes the `image_load_addr`, which represents
     /// the virtual address where the ELF image is loaded in memory. The iterator yields
-    /// [`Elf64ImageLoadSegment`] instances.
+    /// [`load_segments::Elf64ImageLoadSegment`] instances.
     ///
     /// # Arguments
     ///
@@ -666,116 +666,6 @@ impl<'a> Elf64File<'a> {
         self.elf_hdr
             .e_entry
             .wrapping_add(self.load_base(image_load_addr))
-    }
-}
-
-/// Represents a collection of ELF64 load segments, each associated with an
-/// address range and a program header index.
-#[derive(Debug, Default, PartialEq)]
-struct Elf64LoadSegments {
-    segments: Vec<(Elf64AddrRange, Elf64Half)>,
-}
-
-impl Elf64LoadSegments {
-    /// Creates a new empty [`Elf64LoadSegments`] instance.
-    ///
-    /// # Returns
-    /// Returns a new [`Elf64LoadSegments`] with no segments.
-    fn new() -> Self {
-        Self {
-            segments: Vec::new(),
-        }
-    }
-
-    /// Finds the index of the first load segment whose address range does not come before
-    /// the specified `range`.
-    ///
-    /// # Parameters
-    /// - `range`: An [`Elf64AddrRange`] representing the address range to compare against.
-    ///
-    /// # Returns
-    /// Returns [`Some(index)`] if a matching segment is found, where `index` is the index
-    /// of the first such segment. Returns [`None`] if no matching segment is found.
-    fn find_first_not_before(&self, range: &Elf64AddrRange) -> Option<usize> {
-        let i = self.segments.partition_point(|segment| {
-            matches!(segment.0.partial_cmp(range), Some(cmp::Ordering::Less))
-        });
-
-        if i != self.segments.len() {
-            Some(i)
-        } else {
-            None
-        }
-    }
-
-    /// Attempts to insert a new load segment into the collection.
-    ///
-    /// If the segment does not overlap with any existing segments, it is inserted
-    /// into the collection.
-    ///
-    /// # Parameters
-    /// - `segment`: An [`Elf64AddrRange`] representing the address range of the segment to insert.
-    /// - `phdr_index`: An [`Elf64Half`] representing the program header index associated with
-    ///   the segment.
-    ///
-    /// # Returns
-    /// Returns [`Ok`] if the insertion is successful and there is no overlap with existing
-    fn try_insert(&mut self, segment: Elf64AddrRange, phdr_index: Elf64Half) -> Result<(), ()> {
-        let i = self.find_first_not_before(&segment);
-        match i {
-            Some(i) => {
-                match segment.partial_cmp(&self.segments[i].0) {
-                    Some(cmp::Ordering::Less) => {
-                        // Ok, no overlap.
-                        self.segments.insert(i, (segment, phdr_index));
-                        Ok(())
-                    }
-                    _ => Err(()),
-                }
-            }
-            None => {
-                self.segments.push((segment, phdr_index));
-                Ok(())
-            }
-        }
-    }
-
-    /// Looks up an address range and returns the associated program header index and offset
-    /// within the segment.
-    ///
-    /// # Parameters
-    /// - `range`: An [`Elf64AddrRange`] representing the address range to look up.
-    ///
-    /// # Returns
-    /// Returns [`Some((phdr_index, offset))`] if the address range is found within a segment,
-    /// where `phdr_index` is the program header index, and `offset` is the offset within
-    fn lookup_vaddr_range(&self, range: &Elf64AddrRange) -> Option<(Elf64Half, Elf64Xword)> {
-        let i = self.find_first_not_before(range);
-        let i = match i {
-            Some(i) => i,
-            None => return None,
-        };
-
-        let segment = &self.segments[i];
-        if segment.0.vaddr_begin <= range.vaddr_begin && range.vaddr_end <= segment.0.vaddr_end {
-            let offset_in_segment = range.vaddr_begin - segment.0.vaddr_begin;
-            Some((segment.1, offset_in_segment))
-        } else {
-            None
-        }
-    }
-
-    /// Computes the total virtual address range covered by all load segments.
-    ///
-    /// # Returns
-    /// Returns an [`Elf64AddrRange`] representing the total virtual address range covered by
-    /// all load segments. If there are no segments, it returns a range with both boundaries set
-    /// to 0.
-    fn total_vaddr_range(&self) -> Elf64AddrRange {
-        Elf64AddrRange {
-            vaddr_begin: self.segments.first().map_or(0, |first| first.0.vaddr_begin),
-            vaddr_end: self.segments.last().map_or(0, |last| last.0.vaddr_end),
-        }
     }
 }
 
@@ -1011,75 +901,6 @@ impl Elf64Dynamic {
     /// Returns `true` if the PIE flag (DF_PIE_1) is set; otherwise, returns `false`.
     fn is_pie(&self) -> bool {
         self.flags_1 & Self::DF_PIE_1 != 0
-    }
-}
-
-/// Information about the allocation of a virtual address range
-#[derive(Clone, Copy, Debug)]
-pub struct Elf64ImageLoadVaddrAllocInfo {
-    /// The virtual address (vaddr) range to allocate
-    pub range: Elf64AddrRange,
-    /// Optional alignment value set for PIE (Position-Independent
-    /// Executable) executables, allowing a valid vaddr base to be allocated
-    pub align: Option<Elf64Xword>,
-}
-
-/// Represents an ELF64 image load segment
-#[derive(Debug)]
-pub struct Elf64ImageLoadSegment<'a> {
-    /// The virtual address (vaddr) range covering by this segment
-    pub vaddr_range: Elf64AddrRange,
-    /// The contents of the segment in the ELF file
-    pub file_contents: &'a [u8],
-    /// Flags associated with this segment
-    pub flags: Elf64PhdrFlags,
-}
-
-/// An iterator over ELF64 image load segments within an ELF file
-#[derive(Debug)]
-pub struct Elf64ImageLoadSegmentIterator<'a> {
-    elf_file: &'a Elf64File<'a>,
-    load_base: Elf64Xword,
-
-    next: usize,
-}
-
-impl<'a> Iterator for Elf64ImageLoadSegmentIterator<'a> {
-    type Item = Elf64ImageLoadSegment<'a>;
-
-    /// Advances the iterator to the next ELF64 image load segment and returns it.
-    ///
-    /// # Returns
-    ///
-    /// - [`Some<Elf64ImageLoadSegment>`] if there are more segments to iterate over.
-    /// - [`None`] if all segments have been processed.
-    fn next(&mut self) -> Option<Self::Item> {
-        let cur = self.next;
-        if cur == self.elf_file.load_segments.segments.len() {
-            return None;
-        }
-        self.next += 1;
-
-        // Retrieve the program header (phdr) associated with the current segment
-        let phdr_index = self.elf_file.load_segments.segments[cur].1;
-        let phdr = self.elf_file.read_phdr(phdr_index);
-
-        // Calculate the virtual address (vaddr) range based on the phdr information and load base
-        let mut vaddr_range = phdr.vaddr_range();
-        vaddr_range.vaddr_begin = vaddr_range.vaddr_begin.wrapping_add(self.load_base);
-        vaddr_range.vaddr_end = vaddr_range.vaddr_end.wrapping_add(self.load_base);
-
-        // Retrieve the file range for this phdr
-        let file_range = phdr.file_range();
-        // Extract the segment's file contents from the ELF file buffer
-        let file_contents =
-            &self.elf_file.elf_file_buf[file_range.offset_begin..file_range.offset_end];
-
-        Some(Elf64ImageLoadSegment {
-            vaddr_range,
-            file_contents,
-            flags: phdr.p_flags,
-        })
     }
 }
